pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "videosum"
version = "0.1.0"
description = "Small-scale video summarization library with pluggable AI providers"
requires-python = ">=3.10"
dependencies = [
    "anthropic>=0.40.0",
    "ffmpeg-python>=0.2.0",
    "Pillow>=10.0.0",
    "faster-whisper>=1.0.0",
    "httpx>=0.27.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-mock>=3.12",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["videosum*"]
